% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murp.R
\name{murp_specific}
\alias{murp_specific}
\title{BicSpan}
\usage{
murp_specific(
  Data = NULL,
  h = NULL,
  d = NULL,
  cores = NULL,
  omega = NULL,
  K = NULL,
  seed = 723,
  cluster_iter_max = 1000
)
}
\arguments{
\item{Data}{expression matrix with Cells * Genes}

\item{h}{hclust object of orig data, which is not needed without hierarchical clustering}

\item{d}{dist matrix of orig data}

\item{cores}{the number of cores to use in multi-threads calculating,
default is 1. I do not recommend using multi-threaded calculations on
a local computer, but if you have server support, you can try.}

\item{omega}{omega value in pseudo-BIC calculating}

\item{K}{the number of specified clusters}

\item{seed}{random seed, default is 723}
}
\value{
a list include recommend_K, object of clustering result and so on
}
\description{
Calculate the pseu-BIC value under different K.
Loglike: Prior probability estimated by likelyhood function;
MURP: Bayesian Information Criterions value;
Recommended_K: Recommended K value.
}
