% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murp.R
\name{MURP}
\alias{MURP}
\title{MURP}
\usage{
MURP(
  Data,
  cores = 1,
  iter = 20,
  omega = 1/6,
  seed = 723,
  fast = FALSE,
  cluster_iter_max = 1000
)
}
\arguments{
\item{Data}{expression matrix with Cells * Genes}

\item{cores}{the number of cores to use in multi-threads calculating, default is 1.
Multi-threaded calculation is not recommended except for users with server support}

\item{iter}{the number of iterations}

\item{omega}{omega value in pseudo-BIC calculating}

\item{seed}{random seed, default is 723}
}
\value{
a list include recommend_K, object of clustering result and so on
}
\description{
This new method is optimized from the MURP algorithm, it's more efficient.
Loglike: Prior probability estimated by likelyhood function;
MURP: Bayesian Information Criterions value;
Recommended_K: Recommended K value.
}
